completed

+++ controller/ContactController.js
@@ -49,6 +49,12 @@ const updateContact = asyncHandler(async (req, res) => {
     res.status(404);
     throw new Error("Contact not found");
   }
+
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(401);
+    throw new Error("You don't have permission to update this contact");
+  }
+
   const updatedContact = await Contact.findByIdAndUpdate(
     req.params.id,
     req.body,
@@ -66,6 +72,12 @@ const deleteContact = asyncHandler(async (req, res) => {
     res.status(404);
     throw new Error("Contact not found");
   }
+
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(401);
+    throw new Error("You don't have permission to delete this contact");
+  }
+
   const deletedContact = await Contact.findByIdAndDelete(req.params.id);
   res.status(200).json(deletedContact);
 });
